+++ NEWS.md
@@ -9,6 +9,9 @@
 ### folio-spring-kafka
 * [FOLSPRINGS-232](https://folio-org.atlassian.net/browse/FOLSPRINGS-232) Add tenant-aware Kafka message filtering support for Spring modules
 
+### folio-spring-kafka-filtering
+* [FOLSPRINGS-232](https://folio-org.atlassian.net/browse/FOLSPRINGS-232) Add tenant-aware Kafka message filtering support for Spring modules
+
 ## 10.0.0 2026-04-06
 
 **Breaking Changes:**